refactored debouncer
