+++ client/src/hooks/Debouncer/useDebounce.tsx
@@ -1,18 +1,15 @@
-import {useCallback, useEffect, useState} from "react";
-import { UseDebounceProps } from "../../types/Debouncer/types";
-export default function useDebounce<T>({ object, delay = 1000}: UseDebounceProps<T>){
+import { useCallback, useEffect, useState } from "react";
+import { UseDebounceProps } from "@/hooks/Debouncer/types";
+
+export default function useDebounce<T>({ object, delay = 1000 }: UseDebounceProps<T>) {
   const [value, setValue] = useState<T | undefined>()
   const [isBouncing, setBouncing] = useState(true)
 
   const computeDelay = useCallback(() => {
-    if(!object){
-      return 0
-    }
     return delay
   }, [delay, object]);
 
   useEffect(() => {
-
     setBouncing(true)
     const timeout = setTimeout(() => {
       setValue(object)
